+++ age/CHANGELOG.md
@@ -12,6 +12,13 @@ to 1.0.0 are beta releases.
 ### Added
 - `age::armor::ArmoredReadError`, used to wrap armor-specific read errors inside
   `std::io::Error`.
+- `age::ssh`:
+  - `impl Clone for Identity`
+
+### Changed
+- `age::Encryptor::with_recipients` now returns `Option<Encryptor>`, with `None`
+  returned if the provided list of recipients is empty (to prevent files being
+  encrypted to no recipients).
 
 ### Fixed
 - `age::Decryptor` now rejects invalid or non-canonical `scrypt` recipient